Support 4.0.0 features

+++ Database/EventStore/Internal/Operation/DeleteStream/Message.hs
@@ -28,7 +28,7 @@ import Database.EventStore.Internal.Prelude
 data Request
     = Request
       { _streamId        :: Required 1 (Value Text)
-      , _expectedVersion :: Required 2 (Value Int32)
+      , _expectedVersion :: Required 2 (Value Int64)
       , _requireMaster   :: Required 3 (Value Bool)
       , _hardDelete      :: Optional 4 (Value Bool)
       }
@@ -39,7 +39,7 @@ instance Encode Request
 
 --------------------------------------------------------------------------------
 -- | 'Request' smart constructor.
-newRequest :: Text -> Int32 -> Bool -> Maybe Bool -> Request
+newRequest :: Text -> Int64 -> Bool -> Maybe Bool -> Request
 newRequest stream_id exp_ver req_master hard_delete =
     Request
     { _streamId        = putField stream_id
